move TT code to separate file
